[chore] Refactor pdatagen field generators into separate files
