+++ internal/cmd/pdatagen/internal/field.go
@@ -1,0 +1,28 @@
+// Copyright The OpenTelemetry Authors
+// SPDX-License-Identifier: Apache-2.0
+
+package internal // import "go.opentelemetry.io/collector/internal/cmd/pdatagen/internal"
+
+type Field interface {
+	GenerateAccessors(ms *messageStruct) string
+
+	GenerateAccessorsTest(ms *messageStruct) string
+
+	GenerateSetWithTestValue(ms *messageStruct) string
+
+	GenerateCopyOrig(ms *messageStruct) string
+}
+
+func origAccessor(packageName string) string {
+	if usedByOtherDataTypes(packageName) {
+		return "getOrig()"
+	}
+	return "orig"
+}
+
+func stateAccessor(packageName string) string {
+	if usedByOtherDataTypes(packageName) {
+		return "getState()"
+	}
+	return "state"
+}
