+++ internal/cmd/pdatagen/internal/plog_package.go
@@ -12,6 +12,7 @@ var plog = &Package{
 			``,
 			`"go.opentelemetry.io/collector/pdata/internal"`,
 			`"go.opentelemetry.io/collector/pdata/internal/data"`,
+			`"go.opentelemetry.io/collector/pdata/internal/json"`,
 			`otlplogs "go.opentelemetry.io/collector/pdata/internal/data/protogen/logs/v1"`,
 			`"go.opentelemetry.io/collector/pdata/pcommon"`,
 		},
@@ -23,6 +24,7 @@ var plog = &Package{
 			``,
 			`"go.opentelemetry.io/collector/pdata/internal"`,
 			`"go.opentelemetry.io/collector/pdata/internal/data"`,
+			`"go.opentelemetry.io/collector/pdata/internal/json"`,
 			`otlplogs "go.opentelemetry.io/collector/pdata/internal/data/protogen/logs/v1"`,
 			`"go.opentelemetry.io/collector/pdata/pcommon"`,
 		},
@@ -46,10 +48,10 @@ var resourceLogs = &messageStruct{
 	structName:     "ResourceLogs",
 	description:    "// ResourceLogs is a collection of logs from a Resource.",
 	originFullName: "otlplogs.ResourceLogs",
-	fields: []baseField{
+	fields: []Field{
 		resourceField,
 		schemaURLField,
-		&sliceField{
+		&SliceField{
 			fieldName:   "ScopeLogs",
 			returnSlice: scopeLogsSlice,
 		},
@@ -65,10 +67,10 @@ var scopeLogs = &messageStruct{
 	structName:     "ScopeLogs",
 	description:    "// ScopeLogs is a collection of logs from a LibraryInstrumentation.",
 	originFullName: "otlplogs.ScopeLogs",
-	fields: []baseField{
+	fields: []Field{
 		scopeField,
 		schemaURLField,
-		&sliceField{
+		&SliceField{
 			fieldName:   "LogRecords",
 			returnSlice: logSlice,
 		},
@@ -84,45 +86,46 @@ var logRecord = &messageStruct{
 	structName:     "LogRecord",
 	description:    "// LogRecord are experimental implementation of OpenTelemetry Log Data Model.\n",
 	originFullName: "otlplogs.LogRecord",
-	fields: []baseField{
-		&primitiveTypedField{
+	fields: []Field{
+		&TypedField{
 			fieldName:       "ObservedTimestamp",
 			originFieldName: "ObservedTimeUnixNano",
 			returnType:      timestampType,
 		},
-		&primitiveTypedField{
+		&TypedField{
 			fieldName:       "Timestamp",
 			originFieldName: "TimeUnixNano",
 			returnType:      timestampType,
 		},
 		traceIDField,
 		spanIDField,
-		&primitiveTypedField{
+		&TypedField{
 			fieldName: "Flags",
-			returnType: &primitiveType{
+			returnType: &TypedType{
 				structName: "LogRecordFlags",
 				rawType:    "uint32",
 				defaultVal: "0",
 				testVal:    "1",
 			},
 		},
-		&primitiveField{
+		&PrimitiveField{
 			fieldName:  "EventName",
 			returnType: "string",
 			defaultVal: `""`,
 			testVal:    `""`,
 		},
-		&primitiveField{
+		&PrimitiveField{
 			fieldName:  "SeverityText",
 			returnType: "string",
 			defaultVal: `""`,
 			testVal:    `"INFO"`,
 		},
-		&primitiveTypedField{
+		&TypedField{
 			fieldName: "SeverityNumber",
-			returnType: &primitiveType{
+			returnType: &TypedType{
 				structName: "SeverityNumber",
 				rawType:    "otlplogs.SeverityNumber",
+				isType:     true,
 				defaultVal: `otlplogs.SeverityNumber(0)`,
 				testVal:    `otlplogs.SeverityNumber(5)`,
 			},
@@ -133,7 +136,7 @@ var logRecord = &messageStruct{
 	},
 }
 
-var bodyField = &messageField{
+var bodyField = &MessageField{
 	fieldName:     "Body",
 	returnMessage: anyValue,
 }
