+++ internal/cmd/pdatagen/internal/pmetric_package.go
@@ -12,6 +12,7 @@ var pmetric = &Package{
 			``,
 			`"go.opentelemetry.io/collector/pdata/internal"`,
 			`"go.opentelemetry.io/collector/pdata/internal/data"`,
+			`"go.opentelemetry.io/collector/pdata/internal/json"`,
 			`otlpmetrics "go.opentelemetry.io/collector/pdata/internal/data/protogen/metrics/v1"`,
 			`"go.opentelemetry.io/collector/pdata/pcommon"`,
 		},
@@ -23,6 +24,7 @@ var pmetric = &Package{
 			``,
 			`"go.opentelemetry.io/collector/pdata/internal"`,
 			`"go.opentelemetry.io/collector/pdata/internal/data"`,
+			`"go.opentelemetry.io/collector/pdata/internal/json"`,
 			`otlpmetrics "go.opentelemetry.io/collector/pdata/internal/data/protogen/metrics/v1"`,
 			`"go.opentelemetry.io/collector/pdata/pcommon"`,
 		},
@@ -64,10 +66,10 @@ var resourceMetrics = &messageStruct{
 	structName:     "ResourceMetrics",
 	description:    "// ResourceMetrics is a collection of metrics from a Resource.",
 	originFullName: "otlpmetrics.ResourceMetrics",
-	fields: []baseField{
+	fields: []Field{
 		resourceField,
 		schemaURLField,
-		&sliceField{
+		&SliceField{
 			fieldName:   "ScopeMetrics",
 			returnSlice: scopeMetricsSlice,
 		},
@@ -83,10 +85,10 @@ var scopeMetrics = &messageStruct{
 	structName:     "ScopeMetrics",
 	description:    "// ScopeMetrics is a collection of metrics from a LibraryInstrumentation.",
 	originFullName: "otlpmetrics.ScopeMetrics",
-	fields: []baseField{
+	fields: []Field{
 		scopeField,
 		schemaURLField,
-		&sliceField{
+		&SliceField{
 			fieldName:   "Metrics",
 			returnSlice: metricSlice,
 		},
@@ -103,51 +105,51 @@ var metric = &messageStruct{
 	description: "// Metric represents one metric as a collection of datapoints.\n" +
 		"// See Metric definition in OTLP: https://github.com/open-telemetry/opentelemetry-proto/blob/main/opentelemetry/proto/metrics/v1/metrics.proto",
 	originFullName: "otlpmetrics.Metric",
-	fields: []baseField{
+	fields: []Field{
 		nameField,
-		&primitiveField{
+		&PrimitiveField{
 			fieldName:  "Description",
 			returnType: "string",
 			defaultVal: `""`,
 			testVal:    `"test_description"`,
 		},
-		&primitiveField{
+		&PrimitiveField{
 			fieldName:  "Unit",
 			returnType: "string",
 			defaultVal: `""`,
 			testVal:    `"1"`,
 		},
-		&sliceField{
+		&SliceField{
 			fieldName:   "Metadata",
 			returnSlice: mapStruct,
 		},
-		&oneOfField{
+		&OneOfField{
 			typeName:                   "MetricType",
 			originFieldName:            "Data",
 			testValueIdx:               1, // Sum
 			omitOriginFieldNameInNames: true,
 			values: []oneOfValue{
-				&oneOfMessageValue{
+				&OneOfMessageValue{
 					fieldName:              "Gauge",
 					originFieldPackageName: "otlpmetrics",
 					returnMessage:          gauge,
 				},
-				&oneOfMessageValue{
+				&OneOfMessageValue{
 					fieldName:              "Sum",
 					originFieldPackageName: "otlpmetrics",
 					returnMessage:          sum,
 				},
-				&oneOfMessageValue{
+				&OneOfMessageValue{
 					fieldName:              "Histogram",
 					originFieldPackageName: "otlpmetrics",
 					returnMessage:          histogram,
 				},
-				&oneOfMessageValue{
+				&OneOfMessageValue{
 					fieldName:              "ExponentialHistogram",
 					originFieldPackageName: "otlpmetrics",
 					returnMessage:          exponentialHistogram,
 				},
-				&oneOfMessageValue{
+				&OneOfMessageValue{
 					fieldName:              "Summary",
 					originFieldPackageName: "otlpmetrics",
 					returnMessage:          summary,
@@ -161,8 +163,8 @@ var gauge = &messageStruct{
 	structName:     "Gauge",
 	description:    "// Gauge represents the type of a numeric metric that always exports the \"current value\" for every data point.",
 	originFullName: "otlpmetrics.Gauge",
-	fields: []baseField{
-		&sliceField{
+	fields: []Field{
+		&SliceField{
 			fieldName:   "DataPoints",
 			returnSlice: numberDataPointSlice,
 		},
@@ -173,10 +175,10 @@ var sum = &messageStruct{
 	structName:     "Sum",
 	description:    "// Sum represents the type of a numeric metric that is calculated as a sum of all reported measurements over a time interval.",
 	originFullName: "otlpmetrics.Sum",
-	fields: []baseField{
+	fields: []Field{
 		aggregationTemporalityField,
 		isMonotonicField,
-		&sliceField{
+		&SliceField{
 			fieldName:   "DataPoints",
 			returnSlice: numberDataPointSlice,
 		},
@@ -187,9 +189,9 @@ var histogram = &messageStruct{
 	structName:     "Histogram",
 	description:    "// Histogram represents the type of a metric that is calculated by aggregating as a Histogram of all reported measurements over a time interval.",
 	originFullName: "otlpmetrics.Histogram",
-	fields: []baseField{
+	fields: []Field{
 		aggregationTemporalityField,
-		&sliceField{
+		&SliceField{
 			fieldName:   "DataPoints",
 			returnSlice: histogramDataPointSlice,
 		},
@@ -201,9 +203,9 @@ var exponentialHistogram = &messageStruct{
 	description: `// ExponentialHistogram represents the type of a metric that is calculated by aggregating
 	// as a ExponentialHistogram of all reported double measurements over a time interval.`,
 	originFullName: "otlpmetrics.ExponentialHistogram",
-	fields: []baseField{
+	fields: []Field{
 		aggregationTemporalityField,
-		&sliceField{
+		&SliceField{
 			fieldName:   "DataPoints",
 			returnSlice: exponentialHistogramDataPointSlice,
 		},
@@ -214,8 +216,8 @@ var summary = &messageStruct{
 	structName:     "Summary",
 	description:    "// Summary represents the type of a metric that is calculated by aggregating as a Summary of all reported double measurements over a time interval.",
 	originFullName: "otlpmetrics.Summary",
-	fields: []baseField{
-		&sliceField{
+	fields: []Field{
+		&SliceField{
 			fieldName:   "DataPoints",
 			returnSlice: summaryDataPointSlice,
 		},
@@ -231,23 +233,23 @@ var numberDataPoint = &messageStruct{
 	structName:     "NumberDataPoint",
 	description:    "// NumberDataPoint is a single data point in a timeseries that describes the time-varying value of a number metric.",
 	originFullName: "otlpmetrics.NumberDataPoint",
-	fields: []baseField{
+	fields: []Field{
 		attributes,
 		startTimeField,
 		timeField,
-		&oneOfField{
+		&OneOfField{
 			typeName:        "NumberDataPointValueType",
 			originFieldName: "Value",
 			testValueIdx:    0, // Double
 			values: []oneOfValue{
-				&oneOfPrimitiveValue{
+				&OneOfPrimitiveValue{
 					fieldName:       "Double",
 					originFieldName: "AsDouble",
 					returnType:      "float64",
 					defaultVal:      "float64(0.0)",
 					testVal:         "float64(17.13)",
 				},
-				&oneOfPrimitiveValue{
+				&OneOfPrimitiveValue{
 					fieldName:       "Int",
 					originFieldName: "AsInt",
 					returnType:      "int64",
@@ -270,7 +272,7 @@ var histogramDataPoint = &messageStruct{
 	structName:     "HistogramDataPoint",
 	description:    "// HistogramDataPoint is a single data point in a timeseries that describes the time-varying values of a Histogram of values.",
 	originFullName: "otlpmetrics.HistogramDataPoint",
-	fields: []baseField{
+	fields: []Field{
 		attributes,
 		startTimeField,
 		timeField,
@@ -297,28 +299,28 @@ var exponentialHistogramDataPoint = &messageStruct{
 	// summary statistics for a population of values, it may optionally contain the
 	// distribution of those values across a set of buckets.`,
 	originFullName: "otlpmetrics.ExponentialHistogramDataPoint",
-	fields: []baseField{
+	fields: []Field{
 		attributes,
 		startTimeField,
 		timeField,
 		countField,
-		&primitiveField{
+		&PrimitiveField{
 			fieldName:  "Scale",
 			returnType: "int32",
 			defaultVal: "int32(0)",
 			testVal:    "int32(4)",
 		},
-		&primitiveField{
+		&PrimitiveField{
 			fieldName:  "ZeroCount",
 			returnType: "uint64",
 			defaultVal: "uint64(0)",
 			testVal:    "uint64(201)",
 		},
-		&messageField{
+		&MessageField{
 			fieldName:     "Positive",
 			returnMessage: bucketsValues,
 		},
-		&messageField{
+		&MessageField{
 			fieldName:     "Negative",
 			returnMessage: bucketsValues,
 		},
@@ -327,7 +329,7 @@ var exponentialHistogramDataPoint = &messageStruct{
 		sumField,
 		minField,
 		maxField,
-		&primitiveField{
+		&PrimitiveField{
 			fieldName:  "ZeroThreshold",
 			returnType: "float64",
 			defaultVal: "float64(0.0)",
@@ -340,8 +342,8 @@ var bucketsValues = &messageStruct{
 	structName:     "ExponentialHistogramDataPointBuckets",
 	description:    "// ExponentialHistogramDataPointBuckets are a set of bucket counts, encoded in a contiguous array of counts.",
 	originFullName: "otlpmetrics.ExponentialHistogramDataPoint_Buckets",
-	fields: []baseField{
-		&primitiveField{
+	fields: []Field{
+		&PrimitiveField{
 			fieldName:  "Offset",
 			returnType: "int32",
 			defaultVal: "int32(0)",
@@ -360,13 +362,13 @@ var summaryDataPoint = &messageStruct{
 	structName:     "SummaryDataPoint",
 	description:    "// SummaryDataPoint is a single data point in a timeseries that describes the time-varying values of a Summary of double values.",
 	originFullName: "otlpmetrics.SummaryDataPoint",
-	fields: []baseField{
+	fields: []Field{
 		attributes,
 		startTimeField,
 		timeField,
 		countField,
 		doubleSumField,
-		&sliceField{
+		&SliceField{
 			fieldName:   "QuantileValues",
 			returnSlice: quantileValuesSlice,
 		},
@@ -383,7 +385,7 @@ var quantileValues = &messageStruct{
 	structName:     "SummaryDataPointValueAtQuantile",
 	description:    "// SummaryDataPointValueAtQuantile is a quantile value within a Summary data point.",
 	originFullName: "otlpmetrics.SummaryDataPoint_ValueAtQuantile",
-	fields: []baseField{
+	fields: []Field{
 		quantileField,
 		valueFloat64Field,
 	},
@@ -401,21 +403,21 @@ var exemplar = &messageStruct{
 		"// for example the span and trace ID of the active span when the exemplar was recorded.",
 
 	originFullName: "otlpmetrics.Exemplar",
-	fields: []baseField{
+	fields: []Field{
 		timeField,
-		&oneOfField{
+		&OneOfField{
 			typeName:        "ExemplarValueType",
 			originFieldName: "Value",
 			testValueIdx:    1, // Int
 			values: []oneOfValue{
-				&oneOfPrimitiveValue{
+				&OneOfPrimitiveValue{
 					fieldName:       "Double",
 					originFieldName: "AsDouble",
 					returnType:      "float64",
 					defaultVal:      "float64(0.0)",
 					testVal:         "float64(17.13)",
 				},
-				&oneOfPrimitiveValue{
+				&OneOfPrimitiveValue{
 					fieldName:       "Int",
 					originFieldName: "AsInt",
 					returnType:      "int64",
@@ -424,7 +426,7 @@ var exemplar = &messageStruct{
 				},
 			},
 		},
-		&sliceField{
+		&SliceField{
 			fieldName:   "FilteredAttributes",
 			returnSlice: mapStruct,
 		},
@@ -433,9 +435,9 @@ var exemplar = &messageStruct{
 	},
 }
 
-var dataPointFlagsField = &primitiveTypedField{
+var dataPointFlagsField = &TypedField{
 	fieldName: "Flags",
-	returnType: &primitiveType{
+	returnType: &TypedType{
 		structName: "DataPointFlags",
 		rawType:    "uint32",
 		defaultVal: "0",
@@ -443,33 +445,33 @@ var dataPointFlagsField = &primitiveTypedField{
 	},
 }
 
-var exemplarsField = &sliceField{
+var exemplarsField = &SliceField{
 	fieldName:   "Exemplars",
 	returnSlice: exemplarSlice,
 }
 
-var countField = &primitiveField{
+var countField = &PrimitiveField{
 	fieldName:  "Count",
 	returnType: "uint64",
 	defaultVal: "uint64(0)",
 	testVal:    "uint64(17)",
 }
 
-var doubleSumField = &primitiveField{
+var doubleSumField = &PrimitiveField{
 	fieldName:  "Sum",
 	returnType: "float64",
 	defaultVal: "float64(0.0)",
 	testVal:    "float64(17.13)",
 }
 
-var valueFloat64Field = &primitiveField{
+var valueFloat64Field = &PrimitiveField{
 	fieldName:  "Value",
 	returnType: "float64",
 	defaultVal: "float64(0.0)",
 	testVal:    "float64(17.13)",
 }
 
-var bucketCountsField = &primitiveSliceField{
+var bucketCountsField = &PrimitiveSliceField{
 	fieldName:         "BucketCounts",
 	returnType:        "UInt64Slice",
 	returnPackageName: "pcommon",
@@ -478,7 +480,7 @@ var bucketCountsField = &primitiveSliceField{
 	testVal:           "[]uint64{1, 2, 3}",
 }
 
-var explicitBoundsField = &primitiveSliceField{
+var explicitBoundsField = &PrimitiveSliceField{
 	fieldName:         "ExplicitBounds",
 	returnType:        "Float64Slice",
 	returnPackageName: "pcommon",
@@ -487,45 +489,46 @@ var explicitBoundsField = &primitiveSliceField{
 	testVal:           "[]float64{1, 2, 3}",
 }
 
-var quantileField = &primitiveField{
+var quantileField = &PrimitiveField{
 	fieldName:  "Quantile",
 	returnType: "float64",
 	defaultVal: "float64(0.0)",
 	testVal:    "float64(17.13)",
 }
 
-var isMonotonicField = &primitiveField{
+var isMonotonicField = &PrimitiveField{
 	fieldName:  "IsMonotonic",
 	returnType: "bool",
 	defaultVal: "false",
 	testVal:    "true",
 }
 
-var aggregationTemporalityField = &primitiveTypedField{
+var aggregationTemporalityField = &TypedField{
 	fieldName: "AggregationTemporality",
-	returnType: &primitiveType{
+	returnType: &TypedType{
 		structName: "AggregationTemporality",
 		rawType:    "otlpmetrics.AggregationTemporality",
+		isType:     true,
 		defaultVal: "otlpmetrics.AggregationTemporality(0)",
 		testVal:    "otlpmetrics.AggregationTemporality(1)",
 	},
 }
 
-var sumField = &optionalPrimitiveValue{
+var sumField = &OptionalPrimitiveField{
 	fieldName:  "Sum",
 	returnType: "float64",
 	defaultVal: "float64(0.0)",
 	testVal:    "float64(17.13)",
 }
 
-var minField = &optionalPrimitiveValue{
+var minField = &OptionalPrimitiveField{
 	fieldName:  "Min",
 	returnType: "float64",
 	defaultVal: "float64(0.0)",
 	testVal:    "float64(9.23)",
 }
 
-var maxField = &optionalPrimitiveValue{
+var maxField = &OptionalPrimitiveField{
 	fieldName:  "Max",
 	returnType: "float64",
 	defaultVal: "float64(0.0)",
