+++ internal/cmd/pdatagen/internal/templates.go
@@ -48,7 +48,7 @@ var (
 )
 
 func parseTemplate(name string, bytes []byte) *template.Template {
-	return template.Must(template.New(name).Parse(string(bytes)))
+	return template.Must(templateNew(name).Parse(string(bytes)))
 }
 
 func executeTemplate(tmpl *template.Template, data any) string {
@@ -57,4 +57,24 @@ func executeTemplate(tmpl *template.Template, data any) string {
 		panic(err)
 	}
 	return sb.String()
+}
+
+func templateNew(name string) *template.Template {
+	return template.New(name).Funcs(template.FuncMap{
+		"upperFirst": upperFirst,
+		"lowerFirst": lowerFirst,
+		"sub":        sub,
+	})
+}
+
+func upperFirst(s string) string {
+	return strings.ToUpper(s[0:1]) + s[1:]
+}
+
+func lowerFirst(s string) string {
+	return strings.ToLower(s[0:1]) + s[1:]
+}
+
+func sub(a, b int) int {
+	return a - b
 }
