+++ internal/cmd/pdatagen/internal/xpdata_entity_package.go
@@ -11,6 +11,7 @@ var xpdataEntity = &Package{
 		path: filepath.Join("xpdata", "entity"),
 		imports: []string{
 			`"go.opentelemetry.io/collector/pdata/internal"`,
+			`"go.opentelemetry.io/collector/pdata/internal/json"`,
 			`"go.opentelemetry.io/collector/pdata/pcommon"`,
 			`otlpcommon "go.opentelemetry.io/collector/pdata/internal/data/protogen/common/v1"`,
 			`otlpresource "go.opentelemetry.io/collector/pdata/internal/data/protogen/resource/v1"`,
@@ -21,6 +22,7 @@ var xpdataEntity = &Package{
 			`"github.com/stretchr/testify/assert"`,
 			``,
 			`"go.opentelemetry.io/collector/pdata/internal"`,
+			`"go.opentelemetry.io/collector/pdata/internal/json"`,
 			`"go.opentelemetry.io/collector/pdata/pcommon"`,
 		},
 	},
@@ -40,19 +42,19 @@ var entityRef = &messageStruct{
 	structName:     "EntityRef",
 	packageName:    "entity",
 	originFullName: "otlpcommon.EntityRef",
-	fields: []baseField{
+	fields: []Field{
 		schemaURLField,
-		&primitiveField{
+		&PrimitiveField{
 			fieldName:  "Type",
 			returnType: "string",
 			defaultVal: `""`,
 			testVal:    `"host"`,
 		},
-		&sliceField{
+		&SliceField{
 			fieldName:   "IdKeys",
 			returnSlice: stringSlice,
 		},
-		&sliceField{
+		&SliceField{
 			fieldName:   "DescriptionKeys",
 			returnSlice: stringSlice,
 		},
